# docs/openapi.provider.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Pawfront.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5199/"
    }
  ],
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "eventCategory",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "isChildFriendly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "amenities",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "title",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/events/trending": {
      "get": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "take",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/events/{eventId}": {
      "get": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/providers/{providerId}/events/banner-image": {
      "post": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/providers/{providerId}/events": {
      "post": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/providers/{providerId}/events/{eventId}": {
      "put": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "providerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/events/{eventId}/views": {
      "post": {
        "tags": [
          "Pawfront.Api"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/events/{eventId}/shares": {
      "post": {
        "tags": [
          "Pawfront.Api"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/events/{eventId}/inquiries": {
      "post": {
        "tags": [
          "Pawfront.Api"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/events/{eventId}/payout-methods": {
      "post": {
        "tags": [
          "EventEndpoints"
        ],
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveEventPayoutMethodsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEventRequest": {
        "required": [
          "eventCategory",
          "isChildFriendly",
          "title",
          "description",
          "bannerImageUrl",
          "amenities",
          "eventType",
          "startDate",
          "endDate",
          "startTime",
          "endTime",
          "isPaid",
          "price",
          "cancellationPolicy",
          "eventLink",
          "physical"
        ],
        "type": "object",
        "properties": {
          "eventCategory": {
            "type": "string"
          },
          "isChildFriendly": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "bannerImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "eventType": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "endTime": {
            "type": "string",
            "format": "time"
          },
          "isPaid": {
            "type": "boolean"
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "cancellationPolicy": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventLink": {
            "type": [
              "null",
              "string"
            ]
          },
          "physical": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PhysicalEventRequest"
              }
            ]
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "OptionalOfboolean": { },
      "OptionalOfDateOnly": { },
      "OptionalOfdecimal": { },
      "OptionalOfIReadOnlyCollectionOfstring": { },
      "OptionalOfPhysicalEventRequest": { },
      "OptionalOfstring": { },
      "OptionalOfTimeOnly": { },
      "PatchEventRequest": {
        "required": [
          "eventCategory",
          "isChildFriendly",
          "title",
          "description",
          "bannerImageUrl",
          "amenities",
          "eventType",
          "startDate",
          "endDate",
          "startTime",
          "endTime",
          "isPaid",
          "price",
          "cancellationPolicy",
          "eventLink",
          "physical"
        ],
        "type": "object",
        "properties": {
          "eventCategory": {
            "$ref": "#/components/schemas/OptionalOfstring"
          },
          "isChildFriendly": {
            "$ref": "#/components/schemas/OptionalOfboolean"
          },
          "title": {
            "$ref": "#/components/schemas/OptionalOfstring"
          },
          "description": {
            "$ref": "#/components/schemas/OptionalOfstring"
          },
          "bannerImageUrl": {
            "$ref": "#/components/schemas/OptionalOfstring"
          },
          "amenities": {
            "$ref": "#/components/schemas/OptionalOfIReadOnlyCollectionOfstring"
          },
          "eventType": {
            "$ref": "#/components/schemas/OptionalOfstring"
          },
          "startDate": {
            "$ref": "#/components/schemas/OptionalOfDateOnly"
          },
          "endDate": {
            "$ref": "#/components/schemas/OptionalOfDateOnly"
          },
          "startTime": {
            "$ref": "#/components/schemas/OptionalOfTimeOnly"
          },
          "endTime": {
            "$ref": "#/components/schemas/OptionalOfTimeOnly"
          },
          "isPaid": {
            "$ref": "#/components/schemas/OptionalOfboolean"
          },
          "price": {
            "$ref": "#/components/schemas/OptionalOfdecimal"
          },
          "cancellationPolicy": {
            "$ref": "#/components/schemas/OptionalOfstring"
          },
          "eventLink": {
            "$ref": "#/components/schemas/OptionalOfstring"
          },
          "physical": {
            "$ref": "#/components/schemas/OptionalOfPhysicalEventRequest"
          }
        }
      },
      "PhysicalEventRequest": {
        "required": [
          "maximumCapacity",
          "location"
        ],
        "type": "object",
        "properties": {
          "maximumCapacity": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "location": {
            "$ref": "#/components/schemas/EventLocationRequest"
          }
        }
      },
...
      "SaveEventPayoutMethodsRequest": {
        "required": [
          "payoutMethods"
        ],
        "type": "object",
        "properties": {
          "payoutMethods": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateEventRequest": {
        "required": [
          "eventCategory",
          "isChildFriendly",
          "title",
          "description",
          "bannerImageUrl",
          "amenities",
          "eventType",
          "startDate",
          "endDate",
          "startTime",
          "endTime",
          "isPaid",
          "price",
          "cancellationPolicy",
          "eventLink",
          "physical"
        ],
        "type": "object",
        "properties": {
          "eventCategory": {
            "type": "string"
          },
          "isChildFriendly": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "bannerImageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "amenities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "eventType": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "startTime": {
            "type": "string",
            "format": "time"
          },
          "endTime": {
            "type": "string",
            "format": "time"
          },
          "isPaid": {
            "type": "boolean"
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "cancellationPolicy": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventLink": {
            "type": [
              "null",
              "string"
            ]
          },
          "physical": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PhysicalEventRequest"
              }
            ]
          }
        }
      },
...
    }
  },
...
}